Shape Signatures of Fish, Glasses, and Bottles

This data set provides exemplary shape signatures of each five fish, bottles, and pairs of glasses. The planar shapes were derived from original artwork obtained from openclipart.org. Column names encode the type of image.

data(complex.shapes)

data(complex.shapes.raw)

Format

complex.shapes is a matrix containing 100 rows and 15 columns obtained from trace.image. complex.shapes.raw contains a list containing 15 three-dimensional matrices and an array of corresponding labels.

References

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of Time Series. Journal of Statistical Software, 67(5), 1--23.

Author

Andreas Brandmaier brandmaier@mpib-berlin.mpg.de